requirements.txt
Django==1.11
django-cors-headers
hamlpy==0.82.2
django-hamlpy==1.1.1
djangorestframework==3.7.7
gunicorn==19.7.1
requests==2.18.4

